Fix injection of partially instantiated templates
This changes the injection of a partially instantiated template
so that it is rebuilt as a new template with the partial instantiation
resolved.

For purposes of injection, all template instantiation must be complete.
The final result of injection, should look relatively "normal" in the
AST. Thus, we cannot maintain partial instantiation information
from another class when injecting a partially instantiated member.
This would relate the two classes in an abnormal way, in turn
resulting in an AST with special characteristics,
which would cause complications for future development.