Add a new File syntax for the top-level structure of the AST.
Implement parsing and basic AST construction for the file. Also update
the semantic processing of lists to flatten the structure when the list
is empty or a singleton. This reduces the size and indirection of the
AST (hopefully correctly).

Start replacing the use of std::vector with llvm::SmallVector where
appropriate. Update other code style issues.