Clean up declaration and postfix expression parsing.
Make it impossible to parse ill-formed kinds of declarations (with the current
syntax), and separate out the parsing of initializers from the main declaration.
Rewrite signature/declarator parsing in terms of postfix expressions.

Limit "applicative forms" to parens, braces, and identifiers for now. We can't
string together arbitrary chains of literals, for example.