Rethink the grammar and parsing of parameter lists.
Replace parameters and arguments with a single stucture that potentially
combines arbitrary parameters and expressions in a single structure, and
allows for grouping of parameters (x, y : int), subject to semantic
analysis.

Remove tuple, array, and block nodes in favor of list and seq nodes. The
tokens and separators attached to these nodes can be used to determine
their "flavor".